ram: k3-ddrss: Add support for DDR in self-refresh
In IO+DDR the DDR is kept in self-refresh while the SoC cores are
powered off completely. During boot the normal initialization routine of
DDR is then slightly different.

A new binding is introduced that references the wakeup configuration of
the chip called 'ti,wkup-conf'. This is used to determine if the DDR is
expected to be in self refresh. Further it is used to deassert
retention.

The driver now selects SYSCON and conditionally SPL_SYSCON config
symbols.

Signed-off-by: Markus Schneider-Pargmann <msp@baylibre.com>